pyproject.toml
[project]
name = "arvo"
version = "0.1.0"
description = "CLI for scaffolding projects and managing cartridges"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # CLI Framework
    "typer>=0.12.0",
    "rich>=13.0",
    
    # Templating & Scaffolding
    "copier>=9.0",
    
    # Configuration & Validation
    "pyyaml>=6.0",
    "pydantic>=2.0",
    
    # File Manipulation
    "tomlkit>=0.12",
    
    # Git Operations
    "gitpython>=3.1",
]

[project.scripts]
arvo = "arvo.cli:main"

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    
    # Code Quality
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "tach>=0.16.0",
    "pre-commit>=4.0.0",

    # Type Stubs
    "types-pyyaml>=6.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/arvo"]

[tool.hatch.build.targets.wheel.shared-data]
"templates" = "arvo/templates"
"cartridges" = "arvo/cartridges"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = ["-v", "--tb=short"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
